+++ client/console.c
@@ -433,7 +433,7 @@ void Con_CenteredPrint (char *text)
 	int		l;
 	char	buffer[1024];
 
-	l = strlen(text);
+	l = (int)strlen(text);
 	l = (con.linewidth-l)/2;
 	if (l < 0)
 		l = 0;
@@ -655,7 +655,7 @@ void Con_DrawConsole (float frac)
 			text = cls.downloadname;
 
 		x = con.linewidth - ((con.linewidth * 7) / 40);
-		y = x - strlen(text) - 8;
+		y = x - ((int)strlen(text)) - 8;
 		i = con.linewidth/3;
 		if (strlen(text) > i) {
 			y = x - i - 11;
@@ -665,7 +665,7 @@ void Con_DrawConsole (float frac)
 		} else
 			strcpy(dlbar, text);
 		strcat(dlbar, ": ");
-		i = strlen(dlbar);
+		i = (int)strlen(dlbar);
 		dlbar[i++] = '\x80';
 		// where's the dot go?
 		if (cls.downloadpercent == 0)
@@ -682,7 +682,7 @@ void Con_DrawConsole (float frac)
 		dlbar[i] = 0;
 
 #if defined (__APPLE__) || defined (MACOSX)
-                dlbarlen = strlen(dlbar);
+        dlbarlen = (int)strlen(dlbar);
 		snprintf(dlbar + dlbarlen, 1024 - dlbarlen, " %02d%%", cls.downloadpercent);
 #else
 		sprintf(dlbar + strlen(dlbar), " %02d%%", cls.downloadpercent);
